[uc] Crank & new input system
I now rely on interrupts for button presses & crank (rotary encoder).
These interrupt service routines (ISRs) are responsible for calling
a InputEventHandler->HandleInputEvent() with the recognized event.
So, in a sense, the state of each screen is updated whenever a new
interrupt arrives and an InputEvent is fired. I'm not sure if this is
a great way to do this, since it adds some coupling between input handling
and state updating.

Another option I considered was using some sort of event queue: interrupts
enqueue events into this queue and the main loop ensures these events are
consumed by some InputEventHandler. This seems more complex and more resource
intensive, but I may revisit this decision. In practice, the simple "async"
interrupt approach seems to work okay.

Debouncing is a big issue when dealing with interrupts. For the mechanical
switch I'm using, it fires somewhere between 5 and 10 times (both on RISING
and FALING edges), whenever I press it once. Solving this was surprisingly
tricky, but I left a comment in the code.

The crank debouncind, surprisingly, lends itself to a super elegant solution
of using a state machine, and debouncing is automatically handled by simple
transitioning to the current state (so no transition). It works super well
so far.