[uc] Mqtt more or less works
But the code is very poorly structured. I'll package it better next.
In my mind, I'm trying to reconcile these two concepts:

- Callbacks (for example in interruct service routines or MQTT "message
  received" callback)
- Not using global variables

On one hand, while I'm not religiously against global variables, I'd like
to avoid it when possible, if for testing alone.

On the other hand, there is something inherently "global" about interrupt
service routines at least - the microcontroller is configured to jump to
a _single_ function address when a particular pin detects an external
interrupt. A function address is not particularly tied to any instance
of an object (`this` is an implicit parameter of member functions), and we
don't have that during the callback.

I'm thinking we can get around these "issues" by passing a callback by value
with a lambda, if the registration function accepts it. Something like:

```C++
main() {
  MyState state;

  register_callback(
    [&state]() {
      mofidy_state(state);
    }
  );
}
```

In our case, we have:
1. For the MQTT callback definition:
```C++
```

This doesn't seem to port very well.

2. For the Arduino ISR registration:
```C++
void attachInterrupt(uint8_t pin, std::function<void(void)> intRoutine, int mode);
```

I will explore using a lambda for this next.