[uc] Introduce functionality for pushing/popping virtual windows into a display
This odd trick works like this:

When a screen renders itselt, it receives a *Display and calls its rendering functions
(for drawing pixels or text). This display instance knows its size, and the screen can
render itself based on that size.
When a screen renders its sub-components (other screens, elements of a list etc), the
screen can _push_ a temporary Window to the Display, so the display reports this new
size to the sub-components. Once a sub-component is done rendering, the parent compoment
can pop that window and resume its rendering.

This is interesting because each component thinks in terms of the space it has available
when it renders itself. Whether that space is real of virtual, from the component's pers-
pective, that is the entire world.